Add tests for my functions, Closes #109
